+++ email/src/account/discover/http.rs
@@ -1,28 +1,31 @@
-//! # Account config discovery
+//! # Account HTTP discovery
 //!
-//! This module contains everything needed to discover account
-//! configuration from a simple email address, based on the
-//! Thunderbird [Autoconfiguration] standard.
-//!
-//! *NOTE: only IMAP and SMTP configurations can be discovered by this
-//! module.*
-//!
-//! [Autoconfiguration]: https://udn.realityripple.com/docs/Mozilla/Thunderbird/Autoconfiguration#Mechanisms
+//! This module contains everything needed to discover account using
+//! HTTP requests.
 
-use anyhow::bail;
-use bytes::{Buf, Bytes};
-use hyper::{body::to_bytes, client::HttpConnector, Client, Uri};
+use hyper::{body::to_bytes, client::HttpConnector, Client, StatusCode, Uri};
 use hyper_rustls::{HttpsConnector, HttpsConnectorBuilder};
+use log::trace;
+use thiserror::Error;
 
 use crate::Result;
 
 use super::config::AutoConfig;
 
-pub struct Http {
+/// Errors related to account HTTP discovery.
+#[derive(Debug, Error)]
+pub enum Error {
+    #[error("cannot get autoconfig from {0}: {1}")]
+    GetConfigError(Uri, StatusCode),
+}
+
+/// Simple HTTP client using rustls connector.
+pub struct HttpClient {
     client: Client<HttpsConnector<HttpConnector>>,
 }
 
-impl Http {
+impl HttpClient {
+    /// Create a new HTTP client using defaults.
     pub fn new() -> Self {
         let client = Client::builder().build(
             HttpsConnectorBuilder::new()
@@ -35,25 +38,22 @@ impl Http {
         Self { client }
     }
 
-    /// Fetches a given url and returns the XML response (if there is one)
-    async fn get(&self, uri: Uri) -> Result<Bytes> {
-        let res = self.client.get(uri).await?;
+    /// Send a GET request to the given URI and try to parse response
+    /// as autoconfig.
+    pub async fn get_config(&self, uri: Uri) -> Result<AutoConfig> {
+        let res = self.client.get(uri.clone()).await?;
 
-        let is_err = !res.status().is_success();
+        let status = res.status();
         let body = to_bytes(res.into_body()).await?;
 
         // If we got an error response we return an error
-        if is_err {
+        if !status.is_success() {
             let err = String::from_utf8_lossy(&body);
-            bail!("{err}")
-        } else {
-            Ok(body.into())
+            trace!("{err}");
+            return Err(Error::GetConfigError(uri.clone(), status).into());
         }
-    }
 
-    pub async fn get_config(&self, uri: Uri) -> Result<AutoConfig> {
-        let bytes = self.get(uri).await?;
-        let config = serde_xml_rs::from_reader(bytes.reader())?;
+        let config = serde_xml_rs::from_reader(body.as_ref())?;
         Ok(config)
     }
 }